+++ src/main/java/think/rpgitems/power/PowerFlame.java
@@ -35,7 +35,7 @@ public class PowerFlame extends Power implements PowerHit {
 
     @Override
     public String displayText(String locale) {
-        return ChatColor.GREEN + String.format(Locale.get("power.flame", locale), (double) burnTime / 20d);
+        return ChatColor.GREEN + String.format(Locale.get("power.flame", locale), burnTime / 20d);
     }
 
     @Override
